+++ src/kiota/KiotaHost.cs
@@ -346,7 +346,7 @@ public static partial class KiotaHost
         manifestOption.AddAlias("-a");
         return manifestOption;
     }
-    [GeneratedRegex(@"^[a-zA-Z_][\w_-]+", RegexOptions.Singleline, 500)]
+    [GeneratedRegex(@"^[a-zA-Z_][\w]*", RegexOptions.Singleline, 500)]
     private static partial Regex classNameRegex();
     [GeneratedRegex(@"^[\w][\w\._-]+", RegexOptions.Singleline, 500)]
     private static partial Regex namespaceNameRegex();