+++ tests/Kiota.Builder.Tests/Writers/TypeScript/CodeFunctionWriterTests.cs
@@ -377,6 +377,54 @@ public sealed class CodeFunctionWriterTests : IDisposable
         Assert.Contains("?? EnumTypeWithOptionObject.SomeOption", result);
     }
     [Fact]
+    public async Task WritesSerializerBodyEnumCollectionAsync()
+    {
+        var parentClass = TestHelper.CreateModelClass(root, "parentClass");
+        TestHelper.AddSerializationPropertiesToModelClass(parentClass);
+        var propName = "propWithDefaultValue";
+        parentClass.AddProperty(new CodeProperty
+        {
+            Name = propName,
+            Kind = CodePropertyKind.Custom,
+            Type = new CodeType
+            {
+                Name = "string",
+            },
+        });
+        var propertyEnum = new CodeEnum
+        {
+            Name = "EnumTypeWithOption",
+            Parent = parentClass,
+        };
+        var enumOption = new CodeEnumOption() { Name = "SomeOption" };
+        propertyEnum.AddOption(enumOption);
+        var codeNamespace = parentClass.Parent as CodeNamespace;
+        codeNamespace.AddEnum(propertyEnum);
+        parentClass.AddProperty(new CodeProperty
+        {
+            Name = "propWithDefaultEnum",
+            DefaultValue = enumOption.Name,
+            Type = new CodeType
+            {
+                TypeDefinition = propertyEnum,
+                CollectionKind = CodeTypeBase.CodeTypeCollectionKind.Array,
+            }
+        });
+
+        await ILanguageRefiner.RefineAsync(new GenerationConfiguration { Language = GenerationLanguage.TypeScript }, root);
+        var serializerFunction = root.FindChildByName<CodeFunction>($"serialize{parentClass.Name.ToFirstCharacterUpperCase()}");
+        Assert.NotNull(serializerFunction);
+        var parentNS = serializerFunction.GetImmediateParentOfType<CodeNamespace>();
+        Assert.NotNull(parentNS);
+        var complexTypeDefinition = root.FindChildByName<CodeInterface>("SomeComplexType");
+        Assert.NotNull(complexTypeDefinition);
+        parentNS.TryAddCodeFile("foo", serializerFunction, parentClass, complexTypeDefinition);
+        writer.Write(serializerFunction);
+        var result = tw.ToString();
+        Assert.Contains("writeCollectionOfEnumValues<EnumTypeWithOption>(\"propWithDefaultEnum\"", result);
+        Assert.Contains("?? [EnumTypeWithOptionObject.SomeOption]", result);
+    }
+    [Fact]
     public async Task WritesInheritedSerializerBodyAsync()
     {
         var generationConfiguration = new GenerationConfiguration { Language = GenerationLanguage.TypeScript };
@@ -1528,6 +1576,46 @@ public sealed class CodeFunctionWriterTests : IDisposable
         var result = tw.ToString();
 
         Assert.Contains("\"property\": n => { parentClass.property = n.getCollectionOfObjectValues<ArrayOfObjects>(createArrayOfObjectsFromDiscriminatorValue) ?? n.getNumberValue() ?? n.getObjectValue<SingleObject>(createSingleObjectFromDiscriminatorValue) ?? n.getStringValue(); }", result);
+    }
+    [Fact]
+    public void WritesByteArrayPropertyDeserialization()
+    {
+        var intfc = new CodeInterface
+        {
+            Name = "SomeInterface",
+            Kind = CodeInterfaceKind.Model,
+            OriginalClass = new CodeClass() { Name = "SomeClass" }
+        };
+        var deserializationMethod = new CodeMethod
+        {
+            Name = "deserialize",
+            ReturnType = new CodeType { Name = "SomeInterface" },
+            Access = AccessModifier.Public,
+            IsAsync = true,
+            IsStatic = true,
+            Kind = CodeMethodKind.Deserializer,
+        };
+        deserializationMethod.AddParameter(new CodeParameter
+        {
+            Name = "model",
+            Type = new CodeType { TypeDefinition = intfc }
+        });
+        intfc.OriginalClass.AddMethod(deserializationMethod);
+        var prop = new CodeProperty
+        {
+            Name = "property",
+            Type = new CodeType { Name = "base64" }
+        };
+        intfc.AddProperty(prop);
+        var function = new CodeFunction(deserializationMethod);
+        var codeFile = new CodeFile
+        {
+            Name = "someFile",
+        };
+        codeFile.AddElements(function);
+        writer.Write(function);
+        var result = tw.ToString();
+        Assert.Contains("\"property\": n => { model.property = n.getByteArrayValue(); }", result, StringComparison.Ordinal);
     }
 }